Add streaming dependencies to unit tests Separate streaming dependencies from basic WS Refactor streaming client code into a separate module
