+++ .github/workflows/unit_tests.yml
@@ -23,7 +23,7 @@ jobs:
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
-          pip install . -r requirements/test_requirements.txt
+          pip install .[streaming] -r requirements/test_requirements.txt
       - name: Run Tests
         run: |
           pytest tests
