+++ setup.py
@@ -74,7 +74,8 @@ setup(
     license='BSD-3-Clause',
     packages=find_packages(),
     install_requires=get_requirements("requirements.txt"),
-    extras_require={"websocket": get_requirements("websocket.txt")},
+    extras_require={"websocket": get_requirements("websocket.txt"),
+                    "steaming": get_requirements("streaming.txt")},
     zip_safe=True,
     classifiers=[
         'Intended Audience :: Developers',
